:recycle: socketLoader 이벤트 분리
기존의 socketLoader의 경우 이벤트가 모두 들어가 있어 코드가 길고 지저분하였음
pipe 함수를 통해 socekt 이벤트를 분리 하였음
