+++ server/src/utils/pipe.ts
@@ -1,0 +1,6 @@
+const pipe = (...fns: Function[]): Function => {
+  return (firstValue: any) =>
+    fns.reduce((prevValue, fn) => fn(prevValue), firstValue);
+};
+
+export default pipe;
